Restrcuture db interaction folder
